Added functionality for saving labels quality data
The labels tagged with the labelling tool mode of the dashboard can
now be saved to dropbox.

However, the buttons to move to the next and previous photo do not
work and were removed: more work on that will be needed. In fact,
to get to the next photo you cannot simply add 1 to the roof_id
counter, because roof_id are not a continuous sequence with step=1
but have "holes"